requirements.txt
apache-airflow-providers-airbyte==2.1.4
dbt-core==1.1.0
dbt-postgres==1.1.0
dbt-bigquery==1.1.0
python-dateutil==2.8.2
pymongo==3.11.0
psycopg2==2.9.3
faker==13.11.1
python-telegram-bot==13.11
pandas==1.4.2
# SQLAlchemy==1.3.24
PyMySQL==1.0.2
tqdm==4.64.0
google-cloud-bigquery==2.34.3
google-cloud-storage==2.3.0
gspread==5.3.2
pyarrow==6.0.1
redis==3.5.3
boto3==1.16.48
pandas-gbq==0.14.1
networkx==2.5
pika==1.2.1
user-agents==2.2.0
slack-sdk==3.18.1
aiohttp==3.8.1
vncorenlp==1.0.3
gensim==4.0.1
gcsfs==2022.7.1
scipy==1.9.0
numpy==1.19.3